+++ SpringBootProject/Backend/HotelBookingManagementSystem/src/main/java/com/example/HotelBookingManagementSystem/repository/RoomRepository.java
@@ -18,7 +18,7 @@ public interface RoomRepository extends JpaRepository<Room, Long> {
     List<Room> findRoomByHotelName(@Param("hotelName") String hotelName);
 
 
-    List<Room> findByHotelId(int hotelId);
+    List<Room> findByHotelId(long hotelId);
 
 
 }
